feat: dockerize

+++ src/lib/server/db/session.ts
@@ -3,16 +3,26 @@ import { encodeBase32LowerCaseNoPadding, encodeHexLowerCase } from '@oslojs/enco
 import { sha256 } from '@oslojs/crypto/sha2';
 import { userTable, type User } from './schema/user';
 import { sessionTable, type Session } from './schema/session';
-import type { Database } from './connect';
+import { DB as db } from './connect';
 
-export function generateSessionToken(db: Database): string {
+export function generateSessionToken(): string {
 	const bytes = new Uint8Array(20);
 	crypto.getRandomValues(bytes);
 	const token = encodeBase32LowerCaseNoPadding(bytes);
 	return token;
 }
 
-export async function createSession(db: Database, token: string, userId: number): Promise<Session> {
+export async function createSession(token: string, userId: number): Promise<Session> {
+	// Assert that token is valid
+	if (token === null || token === undefined || token === '') {
+		throw new Error('Invalid token');
+	}
+
+	// Assert that userId is valid
+	if (userId === null || userId === undefined || userId === 0) {
+		throw new Error('Invalid userId');
+	}
+
 	const sessionId = encodeHexLowerCase(sha256(new TextEncoder().encode(token)));
 	const session: Session = {
 		id: sessionId,
@@ -23,10 +33,12 @@ export async function createSession(db: Database, token: string, userId: number)
 	return session;
 }
 
-export async function validateSessionToken(
-	db: Database,
-	token: string
-): Promise<SessionValidationResult> {
+export async function validateSessionToken(token: string): Promise<SessionValidationResult> {
+	// Assert that token is valid
+	if (token === null || token === undefined || token === '') {
+		throw new Error('Invalid token');
+	}
+
 	const sessionId = encodeHexLowerCase(sha256(new TextEncoder().encode(token)));
 	const result = await db
 		.select({ user: userTable, session: sessionTable })
@@ -53,7 +65,12 @@ export async function validateSessionToken(
 	return { session, user };
 }
 
-export async function invalidateSession(db: Database, sessionId: string): Promise<void> {
+export async function invalidateSession(sessionId: string): Promise<void> {
+	// Assert that sessionId is valid
+	if (sessionId === null || sessionId === undefined || sessionId === '') {
+		throw new Error('Invalid sessionId');
+	}
+
 	await db.delete(sessionTable).where(eq(sessionTable.id, sessionId));
 }
